+++ Libft/ft_memccpy.c
@@ -1,18 +1,31 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memccpy.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: lrosendo <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/02/13 13:58:16 by lrosendo          #+#    #+#             */
+/*   Updated: 2021/02/13 18:17:47 by lrosendo         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-void    *ft_memccpy(void *dest, const void *src, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
-    unsigned int aux;
+	size_t	aux;
 
-    aux = 0;
-    while (*(char*)src != '\0' && aux < n)
-    {
-        *(char*)dest = *(char*)src;
-        dest++;
-        src++;
-        if(*(char*)(dest - 1) == c)
-            return (dest);
-        aux++;
-    }
-    *(char*)dest = 0;
-    return (0);
+	aux = 0;
+	while (*(char*)src != '\0' && aux < n)
+	{
+		*(char*)dest = *(char*)src;
+		dest++;
+		src++;
+		if (*(char*)(dest - 1) == c)
+			return (dest);
+		aux++;
+	}
+	*(char*)dest = 0;
+	return (0);
 }
