+++ Libft/ft_memmove.c
@@ -1,14 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memmove.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: lrosendo <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/02/13 13:58:35 by lrosendo          #+#    #+#             */
+/*   Updated: 2021/02/13 18:19:14 by lrosendo         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-void    *ft_memmove(void *dest, const void *src, unsigned int n)
+#include "libft.h"
+
+void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-    char c;
+	char	c;
 
-    while (n--)
-    {
-        c = *(char*)src;
-        *(char*)dest = c;
-        src++;
-        dest++;
-    }
-    return (dest);
+	while (n--)
+	{
+		c = *(char*)src;
+		*(char*)dest = c;
+		src++;
+		dest++;
+	}
+	return (dest);
 }
