+++ Libft/ft_strlcat.c
@@ -1,33 +1,46 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strlcat.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: lrosendo <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/02/13 13:59:08 by lrosendo          #+#    #+#             */
+/*   Updated: 2021/02/13 18:13:58 by lrosendo         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-unsigned int    ft_len(const char *str)
+#include "libft.h"
+
+size_t	ft_len(const char *str)
 {
-    unsigned int index;
+	size_t	index;
 
-    index = 0;
-    while (str[index])
-        index++;
-    return (index);
+	index = 0;
+	while (str[index])
+		index++;
+	return (index);
 }
 
-unsigned int    ft_strlcat(char *dst, const char *src, unsigned int size)
+size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
-    unsigned int to_return;
-    unsigned int index;
-    unsigned int dst_size;
-    unsigned int src_size;
+	size_t to_return;
+	size_t index;
+	size_t dst_size;
+	size_t src_size;
 
-    dst_size = ft_len(dst);
-    src_size = ft_len(src);
-    index = dst_size;
-    if (dst_size >= size)
-        return (size + src_size);
-    else
-        to_return = dst_size + src_size;
-    while(index < size - 1)
-    {
-        dst[index] = src[index - dst_size];
-        index++;
-    }
-    dst[index] = '\0';
-    return (to_return);
+	dst_size = ft_len(dst);
+	src_size = ft_len(src);
+	index = dst_size;
+	if (dst_size >= size)
+		return (size + src_size);
+	else
+		to_return = dst_size + src_size;
+	while (index < size - 1)
+	{
+		dst[index] = src[index - dst_size];
+		index++;
+	}
+	dst[index] = '\0';
+	return (to_return);
 }
