+++ Libft/ft_strlcpy.c
@@ -1,22 +1,35 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strlcpy.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: lrosendo <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/02/13 13:59:13 by lrosendo          #+#    #+#             */
+/*   Updated: 2021/02/13 18:15:33 by lrosendo         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-unsigned int    ft_strlcpy(char *dst, const char *src, unsigned int size)
+#include "libft.h"
+
+size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-    unsigned int index;
-    unsigned int count;
+	size_t	index;
+	size_t	count;
 
-    count = 0;
-    index = 0;
-    while (src[index])
-    {
-        index++;
-    }
-    if (size == 0)
-        return (index);
-    while (count < (size - 1) && src[count])
-    {
-        dst[count] = src[count];
-        count++;
-    }
-    dst[count] = '\0';
-    return (index);
+	count = 0;
+	index = 0;
+	while (src[index])
+	{
+		index++;
+	}
+	if (size == 0)
+		return (index);
+	while (count < (size - 1) && src[count])
+	{
+		dst[count] = src[count];
+		count++;
+	}
+	dst[count] = '\0';
+	return (index);
 }
