+++ Libft/ft_strlen.c
@@ -1,13 +1,26 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strlen.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: lrosendo <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/02/13 13:59:18 by lrosendo          #+#    #+#             */
+/*   Updated: 2021/02/13 19:14:10 by lrosendo         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-unsigned int    ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-    unsigned int size;
+	size_t	size;
 
-    size = 0;
-    while (*s)
-    {
-        size++;
-        s++;
-    }
-    return (size);
+	size = 0;
+	while (*s)
+	{
+		size++;
+		s++;
+	}
+	return (size);
 }
