+++ Libft/ft_strnstr.c
@@ -1,35 +1,48 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strnstr.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: lrosendo <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/02/13 13:59:26 by lrosendo          #+#    #+#             */
+/*   Updated: 2021/02/13 18:20:20 by lrosendo         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-int     ft_confirm(const char *dest, const char *src, unsigned int aux)
+#include "libft.h"
+
+int		ft_confirm(const char *dest, const char *src, size_t aux)
 {
-    int aux2;
+	int	aux2;
 
-    aux2 = 0;
-    while(src[aux2])
-    {
-        if (src[aux2] != dest[aux])
-            return (0);
-        aux++;
-        aux2++;
-    }
-    return (1);
+	aux2 = 0;
+	while (src[aux2])
+	{
+		if (src[aux2] != dest[aux])
+			return (0);
+		aux++;
+		aux2++;
+	}
+	return (1);
 }
 
-char    *ft_strnstr(const char *dest, const char *src, unsigned int len)
+char	*ft_strnstr(const char *dest, const char *src, size_t len)
 {
-    unsigned int    aux;
-    int             to_return;
+	size_t			aux;
+	int				to_return;
 
-    aux = 0;
-    to_return = 0;
-    while(dest[aux] && aux < len)
-    {
-        if (dest[aux] == src[0])
-        {
-            to_return = ft_confirm(dest, src, aux);
-            if (to_return == 1)
-                return ((char*)(dest + aux));
-        }
-        aux++;
-    }
-    return (0);
+	aux = 0;
+	to_return = 0;
+	while (dest[aux] && aux < len)
+	{
+		if (dest[aux] == src[0])
+		{
+			to_return = ft_confirm(dest, src, aux);
+			if (to_return == 1)
+				return ((char*)(dest + aux));
+		}
+		aux++;
+	}
+	return (0);
 }
